refactor: Move uvicorn integration from _main into new _uvicorn module
ShinyConfig, ShinyServer, _run_uvicorn, maybe_setup_rsw_proxying,
_set_workbench_kwargs, and ReloadArgs now live in shiny/_uvicorn.py.
