+++ shiny/_main.py
@@ -4,14 +4,11 @@ import copy
 import importlib
 import importlib.util
 import inspect
-import logging
 import os
 import platform
 import re
-import socket
 import sys
 import types
-import warnings
 from collections.abc import Callable
 from functools import partial
 from pathlib import Path
@@ -20,16 +17,17 @@ from typing import Any, Optional
 import click
 import uvicorn
 import uvicorn.config
-from uvicorn.config import Config
-from uvicorn.server import Server
-from uvicorn.supervisors import ChangeReload, Multiprocess
 
 import shiny
 
-from . import __version__, _autoreload, _hostenv, _launchbrowser, _static, _utils
+from . import __version__, _autoreload, _launchbrowser, _static, _utils
+from ._uvicorn import (
+    ReloadArgs,
+    _run_uvicorn,
+    _set_workbench_kwargs,
+    maybe_setup_rsw_proxying,
+)
 from ._docstring import no_example
-from ._hostenv import is_workbench
-from ._typing_extensions import NotRequired, TypedDict
 from .bookmark._bookmark_state import shiny_bookmarks_folder_name
 from .express import is_express_app
 from .express._utils import escape_to_var_name
@@ -42,7 +40,6 @@ def main() -> None:
 
 
 stop_shortcut = "Ctrl+C"
-_UVICORN_STARTUP_FAILURE = 3
 
 RELOAD_INCLUDES_DEFAULT = (
     "*.py",
@@ -64,78 +61,6 @@ RELOAD_EXCLUDES_DEFAULT = (
     ".venv",
     shiny_bookmarks_folder_name,
 )
-
-
-class ShinyConfig(Config):
-    # uvicorn's Config assigns these in __init__, but the generated stubs
-    # (`make pyright-typings`) omit instance attributes, so declare them here.
-    reload: bool
-    workers: int
-    uds: str | None
-
-
-class ShinyServer(Server):
-    # uvicorn's Server assigns this in __init__; see ShinyConfig note above.
-    started: bool
-
-    # In reload mode, on_started travels with the server target, so keep it picklable.
-    def __init__(
-        self, config: Config, on_started: Callable[[], None] | None = None
-    ) -> None:
-        super().__init__(config=config)
-        self._on_started = on_started
-
-    async def startup(self, sockets: list[socket.socket] | None = None) -> None:
-        await super().startup(sockets=sockets)
-        if self.started and self._on_started is not None:
-            self._on_started()
-
-
-def _run_uvicorn(
-    app: Any,
-    *,
-    app_dir: str | None = None,
-    on_started: Callable[[], None] | None = None,
-    **kwargs: Any,
-) -> None:
-    # Mirrors uvicorn.run(); re-check this helper when upgrading Uvicorn.
-    if app_dir is not None:
-        sys.path.insert(0, app_dir)
-
-    config = ShinyConfig(app, **kwargs)
-    server = ShinyServer(config=config, on_started=on_started)
-
-    if config.reload or config.workers > 1:
-        if not isinstance(app, str):
-            logging.getLogger("uvicorn.error").warning(
-                "You must pass the application as an import string to enable "
-                "'reload' or 'workers'."
-            )
-            sys.exit(_UVICORN_STARTUP_FAILURE)
-    elif hasattr(config, "load_app"):
-        # Fail fast on app import errors before binding the socket, matching
-        # uvicorn.run(). Config.load_app() only exists in recent Uvicorn releases;
-        # older supported versions (>=0.16) load the app inside serve() instead.
-        config.load_app()
-
-    try:
-        if config.should_reload:
-            sock = config.bind_socket()
-            ChangeReload(config, target=server.run, sockets=[sock]).run()
-        elif config.workers > 1:
-            sock = config.bind_socket()
-            Multiprocess(config, target=server.run, sockets=[sock]).run()
-        else:
-            server.run()
-    except KeyboardInterrupt:
-        pass
-    finally:
-        uds = config.uds
-        if uds and os.path.exists(uds):
-            os.remove(uds)
-
-    if not server.started and not config.should_reload and config.workers == 1:
-        sys.exit(_UVICORN_STARTUP_FAILURE)
 
 
 @main.command(help=f"""Run a Shiny app (press {stop_shortcut} to stop).
@@ -509,17 +434,6 @@ def setup_hot_reload(
     _autoreload.start_server(autoreload_port, app_port, launch_browser)
 
 
-def maybe_setup_rsw_proxying(log_config: dict[str, Any]) -> None:
-    # Replace localhost URLs emitted to the log, with proxied URLs
-    if _hostenv.is_workbench():
-        if "filters" not in log_config:
-            log_config["filters"] = {}
-        log_config["filters"]["rsw_proxy"] = {"()": "shiny._hostenv.ProxyUrlFilter"}
-        if "filters" not in log_config["handlers"]["default"]:
-            log_config["handlers"]["default"]["filters"] = []
-        log_config["handlers"]["default"]["filters"].append("rsw_proxy")
-
-
 def is_file(app: str) -> bool:
     return "/" in app or app.endswith(".py")
 
@@ -777,25 +691,6 @@ def cells_to_app(json_file: str, py_file: str) -> None:
 @main.command(help="""Get Shiny's HTML dependencies as JSON.""")
 def get_shiny_deps() -> None:
     print(shiny.quarto.get_shiny_deps())
-
-
-class ReloadArgs(TypedDict):
-    reload: NotRequired[bool]
-    reload_includes: NotRequired[list[str]]
-    reload_excludes: NotRequired[list[str]]
-    reload_dirs: NotRequired[list[str]]
-
-
-def _set_workbench_kwargs(kwargs: dict[str, Any]) -> None:
-    if is_workbench():
-        if kwargs.get("ws_per_message_deflate"):
-            # Workaround for nginx/uvicorn issue within Workbench
-            # https://github.com/rstudio/rstudio-pro/issues/7368#issuecomment-2918016088
-            warnings.warn(
-                "Overwriting kwarg `ws_per_message_deflate=True` to `False` to avoid breaking issue in Workbench",
-                stacklevel=2,
-            )
-        kwargs["ws_per_message_deflate"] = False
 
 
 # Check that the version of rsconnect supports Shiny Express; can be removed in the
